+++ src/main/java/com/example/demo/service/impl/SendererServiceImpl.java
@@ -1,10 +1,11 @@
-package com.example.demo.service;
+package com.example.demo.service.impl;
 
-import com.example.demo.Entity.Senderer;
+import com.example.demo.entity.Senderer;
 import com.example.demo.dto.SendererDTO;
 import com.example.demo.dto.converter.SendererDtoConverter;
 import com.example.demo.dto.request.CreateSendererRequest;
 import com.example.demo.repo.SendererRepository;
+import com.example.demo.service.SendererService;
 import org.springframework.stereotype.Service;
 
 import java.util.List;
